Catch ValueError in APIError response message.
If the Data API tries to index / delete a service in the
Search API when the Search API isn't running, the Data API returns
a 500-level response without a `.json()` method and everything blows up.
Probably we need to decide what to do if the Data API can't connect
to the Search API, but, whatever we choose, I don't think we want to return
a ValueError.