Deny third-party sites embedding ours in an iframe
Added an `X-Frame-Options` to all headers which prevents any of
our sites being embedded in an iframe by other sites.
Things to consider:
- We *could* set this at the server level if we wanted to make
this change more uniform. (Might be harder to test though)
- The alphagov/whitehall team have set their public-facing headers
for their content to `ALLOWALL`, but @minglis says we're not going
to allow this at present for any of our sites.
https://github.com/alphagov/whitehall/commit/9f8c7ccd1fdf7695e66b1340cb6b89ca0e134f9a