Content loader shouldn't require service attributes
Currently, if the display of a question is dependent on, for example,
the `lot` of a service then `service_data.lot` must be defined.

This isn't a problem in real-world scenarios, because services always have lots.

It is, however, a problem when mocking a `get_service` call for testing.

This commit adds a line to catch any 'services' that don't have set the
parameter that a question depends on.