Split file handling into its own class
Because files are slow to load, the intended use of ContentLoader was to
do this once when starting the app.

This caused thread safety issues when using ContentLoader in an
object-orientated way, because you couldn't reliably store state inside an
instance.

Being able to store state in an instance is nice because it makes the interface
cleaner.

This commit splits content loader into two parts:
- `YAMLLoader`, which is expected to instantiated once on startup and handles
  reading and parsing YAML files, and caching the results
- `ContentBuilder` which is expected to be instantiated once per request
  and therefore could maintain state