Enforce Zulu time on our datetime string.
Python's datetime formatting won't always append a timezone to when
creating a string from a datetime, which then throws an error when
trying to interpret back the string without a timezone.
http://stackoverflow.com/questions/14762518/python-datetime-strptime-and-strftime-how-to-preserve-the-timezone-informat
Instead of using another library, we're going to enforce a 'Z' on the
end of our datetime format string and use `utcnow()` everywhere.