Fix two problems
Problem 1
-------------------------------------------------------------------------------
When sharing one instance of YAMLLoader between two instances of ContentLoader,
filtering persisted, so that any questions removed by one filter were not ever
visible again.

Problem 2
-------------------------------------------------------------------------------
The _populate_section method was replacing strings with dicts in-place, so when
it was run a second time it was expecting strings but found dicts.

--

This commit fixes both of these problems by making copies of objects before
modifying them.