Make YAML loader throw None if file not found
There are places where we, reasonably, do stuff like:
```
for key, value in service_data:
  content.get_question(key)
```

`service_data` has a number of attributes (eg `supplierName`) which don't have
associated content because they're not editable.

Thus it is preferable for `.get_question` to return None, rather than throw
an exception which would have to be caught somewhere further down the line.