Add `create_new_draft` method, rename some others
Data API Client had a `create_service` method which corresponded to
`import_service` in the Data API, and it had a `create_new_draft` method
which, as @TheDoubleK pointed out, was copying an existing service into
a draft.
This new endpoint is creating a new draft service *not associated with
an existing service*, so it felt more appropriate to pilfer an existing
name and try to rename the other one.