Add page iteration to the DataAPIClient
This allows any 'find' method to be called in a way that automatically
iterates through pages and retries failed requests.

Any method that starts with 'find_' and ends with '_iter' it will be
called with an exponential backoff. From the response the records
will be yielded and then the 'next' URL will be fetched and yielded
until there are no pages.

This is most often needed for scripts that export or migrate large
numbers of records where ocassional requests may fail.