Add `get_section_data` to `ContentBuilder`
This moves the form parsing and filtering logic implemented in the
supplier frontend into the `ContentBuilder` so that it can be used on
other content manifests and in other frontends. I have left a couple of
TODOs in around allowing questions with multiple fields (ie. pricing). I
don't want to implement it right now but I would like to when we have
done pre-pag.