Add JSON application logs
This adds a JSON log file when not in development. CloudWatch logs
supports much more sophisticated searching over JSON logs than it does
over plain text logs [1]. Most notably compound conditions (AND and OR).

The JSON log file path is the same as the line by line one with a
.json appended, as it seems simpler than having to explicitly
configure a separate JSON log path.

This uses python-json-logger [2]. I am pulling the v0.1.3 tag directly
from git because there is no v0.1.3 package file in PyPi. We need v0.1.3
for logging excecption track traces. There is an issue open for this
[3].

Some of the keys are rewritten to match the field names used in other
logs such as the apache and nginx access logs.

[1] http://docs.aws.amazon.com/AmazonCloudWatch/latest/DeveloperGuide/FilterAndPatternSyntax.html
[2] https://github.com/madzak/python-json-logger
[3] https://github.com/madzak/python-json-logger/issues/30