Add log message formatting
Allow the log message to be passed in as a format string that is
formatted with the log record. This allows additional parameters to be
passed in once and used for both JSON log fields and the main log
message. This has been implemented for both the JSON logs and the
plain text logs.