Add template filter for chunking the service ID
This adds a template filter for chunking the service ID for display. If
a service ID is all numeric (ie. not G-Cloud 5) it will be split 4-digit
chunks for display.

Currently this chunking is done in a couple of different places; in the
buyer frontend it is done with service_presenters and in the admin app
dmutils/presenters.

This does create a dependency between the frontend toolkit and dmutils.