Fix updating of user's passwords
The API has changed to require who is updating a user:
https://github.com/alphagov/digitalmarketplace-api/commit/7f578ea8a83a29d74f0ac6ff9748f3be93ece89f

This change was reflected in utils for updating a user's information:
https://github.com/alphagov/digitalmarketplace-utils/commit/8a4bd48b1cbd2b6ec682178326ba7ab0ff7a49bf

…but wasn't changed for updating a user's password. This is currently causing
a "Could not update password due to an error" error on production.

This commit allows the email address of the current user to be passed in,
fixing the bug.