Add index argument to SearchAPIClient methods instead of constructor
One SearchAPIClient instance can be used to send requests to multiple
search indexes. Index needs to be set in each method call.

Adds a temprorary default value for the index to keep compatibility
with existing client usage. Once support for g-cloud and dos indexes
is added to the apps and index argument is passed on every method call
the default value can be removed.