Convert integer config values to ints
When a configuration option is defined as an integer it should be
converted to an integer when passed in from the environment. If an
invalid value is passed on the environment then it should fail.

This change also makes boolean configuration options fail if passed a
non-truthy value on the environment.