Support frameworks and manifests in ContentLoader
As we move to handling multiple frameworks the `ContentLoader` needs
to be able to support this. This commit changes the `ContentLoader`'s
role so that it is responsible for managing multiple manifests across
multiple frameworks. Rather than creating multiple `ContentLoader`
instances we create one, pointing it at the root content directory
(`digitalmarketplace-frameworks`). On application startup we load the
manifests we need and then get builders by framework slug and manifest
name. If a manifest is not found either when loading or when getting a
buidler we error with a `ContentNotFound` error and should fail fast.

Example:
```python
loader = ContentLoader('path/to/digitalmarketplace-frameworks')
loader.load_manifest('g-cloud-7', 'declaration', 'declaration')
loader.get_builder('g-cloud-7', 'declaration')
```

Note that this changes `read_yaml` so that it raises an error if the
file is not found. `read_yaml` only seems to be used by the content
loader so this should not cause errors.