Enforce pre-loading of messages
Previously get message could still, successfully, use a file which wasn't
loaded on application startup. This means that the app could be running for
a while until the user hit a page which tried to load the offending file.

This would manifest itself if a developer removed a content file without
checking the page that required it.

This commit changes the behaviour so that any message a developer wants to use
must be explicitly preloaded.