Add support for loading nested questions
Reuses the existing section loading to load `questions:` for question
files. Since the original function also added an `id` to the section
and we need a way to create URLs addressing `multiquestions` we kept
the existing logic but replaced `id` key with `slug` to differentiate
between addressable and non-addressable questions.

Section `slug` is only used internally for now: once section data is
transformed into `ContentSection` instance it can be addressed as
`section.id`. Since renaming it is a breaking change it's kept as
TODO for now. Questions should use `.slug` for URL referencing.