Add ContentQuestionSummary to replace service_attribute
Adds a new class for a question containing a reference to the service
data. This allows adding additional methods to the question: `.value`
and `.answer_required`.

`ContentQuestionSummary` instances are created by calling `.summary`
on manifest or section objects, which returns a copy of the object
with all questions transformed to `ContentQuestionSummary` instances.

This replaces `service_attribute.Attribute` class and removes the need
for creating section look-alike dictionaries with "rows" key for summary
table display since ordinary section objects can be used instead and
there's no need to propagate the changes in section attributes.