Save submission documents separately
With the new bucket layout service submission documents should be saved
separately to service documents. In order for Nginx to properly route
requests the paths must also be different. This change adds a property
to S3 instances that attempts to parse a digitalmarketplace bucket name
and return the short name. This is then used when generating the
document path in S3.