Send empty list value to the API for list/checkbox questions
New JSON schemas always define minItems for checkbox/list questions.
Since questions can be required to have a value even if the field name
is not listed in the schema required keys (e.g. questions that are part
of a multiquestions and have to be answered if any other related
question was answered) we can no longer send empty update to the API and
rely on the API returning an error based on mismatch between required
and request page_questions.

Instead, for list and checkbox questions we send the empty list value to
the API, so that it's validated against the `minItems` requirement.