Use question.label instead of 'question' for field errors list
Error 'question' is displayed in the "there was a problem with the
follwing quesitons" list. Similar to the summary page, it should
prefer the short question.name to the full question text since it
fits better in the field list without taking too much space and
since full question text can contain markdown links which wouldn't
be rendered during the error fields display.