ContentQuestion accepts and properly formats `yes_no_question`s
Supplier responses to briefs are going to be returning a new data type to the `ContentLoader`.
They're `yes_no_questions` in the manifest files -- essentially lists of boolean values.

Like with "yes/no" radio buttons, a list of "true" or "false" values will be returned to
the `ContentLoader` which will then format them as `True` or `False` values for the API.