Add indices to `yes_no_question`s
As @kevkeenoy rightly pointed out, getting back a simple list of booleans
from a submitted form has some important drawbacks: we wouldn't be able to
tell if booleans values have been returned in order (although we assume
this with `list` values), but -- more importantly -- if values are missing,
we wouldn't know which one(s) in the sequence have been omitted.

After failing validation for providing the wrong number of answers, the
values repopulating the form would potentially be filling in answers that the
user hadn't selected if we assumed an order while redoing the form.

The new `yes_no_questions` are to be returned with keys that contain the
`question_id` and an index separated by a hyphen. This means that we can process
returned values by converting correct values to boolean and setting missing values
to None.

Adds a lot more logic to the `_get_single_question_data()` method in the ContentLoader,
but it only applies to this new question type.