Send logs to STDERR only if DM_LOG_PATH is not set
At the moment, the decision to write logs to a file instead of to
stdout/stderr is based on the value of DEBUG flag, with DEBUG=False
environments writing to files and DEBUG=True to stderr. This worked
reasonably well when all non-DEBUG environments were hosted on AWS
and the only DEBUG ones were development and testing. But this split
is less useful for eg deploying to PaaS, since Cloud Foundry expects
logs to be sent to stdout and stderr, but at the same time we don't
necessarily want to switch apps to debug mode.

Basing the logging decision solely on DM_LOG_PATH makes more sense and
would also allow us to switch debug on/off without redirecting the logs.