Add dates package
The dates package contains the logic for working out publishing dates
for briefs, and is currenty used by the buyer frontend and the api.

The way the dates are calculated is slightly different depending on if
the get_publishing_dates function is used by the api or the frontend due
to the way that briefs are serialized when returned to the frontend.

Working days are taken into account via the workdays package brought in
with pip.