Add `SIGNED_SIGNATURE_PAGE_PREFIX`
To mirror the g7 process where we have a framework agreement filename
and then a prefix for when it has been signed. For g8 onwards we
have a signature page filename and then a prefix for when it has
been signed.

We will be able to use this `SIGNED_SIGNATURE_PAGE_PREFIX` in the
front end to set a "pretty" download name for an s3 file.

Have improved the formatting of how we lay out these path constants
so it's clearer how they related to each other.