Allow setting custom disposition-type on S3 saves with filename
For G8 countersigning we want uploaded agreements to have `inline` disposition type but also have a filename set for when users download the files.

(The admin app inline display of PDFs breaks if Content-Disposition is set to `attachment`, but at present we can only set `attachment` with a filename.)

Currently you can only set a custom disposition-type when a filename is also supplied.  This could be changed in future if necessary to set disposition types without a filename.