Add `degenerate_document_path_and_return_doc_name` util
Takes a document path and return just the generic document
name. This will be useful for our frontend apps so they can retrieve
a full document path and strip it just the generic document name.
This document name can then be passed as a param to a
`download_document` route that creates a signed url for that document.

Doing this allows us to keep our download document route generic and
it does not need to have any knowledge of specific agreement keys
such as `agreementPath` and will work for both timestamped files
(such as a signed framework agreement) and non timestamped files
(such as a result letter) in the same way.