Update logging for PaaS environment
When we move to PaaS we want to log json format to stdout. We also need,
in the short term, to log json to file in the Elastic Beanstalk
environment. This is determined by the `DM_LOG_PATH` env variable.

We're also dropping logging both plain text and json logs in production.
Be default we'll use json format, but this can be overwritten by setting
the `DM_PLAIN_TEXT_LOGS` variaable to `True` in the app config.

This will only be done in the dev environment, as seeing the plain text
logs would be more useful than json.