Create DMMailChimpClient class
This functionality used to exist in dm scripts repository as
separate functions. Upon moving them to utils we decided to follow
the pattern used with our other email clients, such as `DMNotifyClient`.
We're moving them to utils as we're going to be adding more
functionality to call the mailchimp api.