Stop checking for question type.
The logic around what to do with different question types
probably shouldn't live here.

Question types are something we have invented inside of our
technical system. This jinja filter should act predictably and
not have to know about what question types it should be operating
on.

Planning to push the knowledge of question types into the app
template rather than here.

Have updated the tests to remove the extra parameter and removed
the test that passes in a different question type.