Decode expired tokens and return minimal data
Currently we don't differentiate between an expired token and a bad
one. Both just return `None`.

It would be more helpful when creating users in the user-frontend
app if we could determine the role that was being attempted to create with
an expired token. That way we can give appropriate advice.

If the token is bad it's probably a copy paste error, and we're less
concerned about giving more helpful advise other than it being a bad
token.