Add `InviteUser` class to emails
The code for emailing an invite email to a buyer and a supplier is
really similar. We're also moving these emails to Notify which also
simplifies the process (no need for templates).

This class can be imported by the frontend apps and used to generate and
send the tokens and emails for user invites.