Add `UserAccountEmail` class to emails
The code for emailing a creation or invite email to a buyer or supplier
is really similar. We're also moving these emails to Notify which also
simplifies the process (no need for templates).

This class can be imported by the frontend apps and used to generate and
send the tokens and emails for user creation/invitiation.