Add milliseconds to the log time format
Allows keeping the order of events when logs are streamed to Kibana.

Removing TIME_FORMAT argument from log formatters makes python use
the default `formatTime` logic that adds `record.msec` to the formatted
string. The default time format is the same as the one we define, except
it uses a space instead of `T` to separate date from time. Since we
don't rely on splitting fields with whitespace in JSON logs this
shouldn't be a problem.