If no "timestamp" metadata fall back to AWS last_modified
Some pages in our apps (eg framework dashboard) look for the existence of "last_modified" on an S3 key as an indicator for if the file exists or not, and only show a link to files with a "last_modified" date.

This meant that if a file was uploaded to S3 using the AWS console and didn't have "timestamp" set then the link to the file would not display, even though the file is there.