Use `SHARED_EMAIL_KEY` not `SECRET_KEY` for reset password tokens
We will be changing our tokens for reset password emails to be
generated using the `SHARED_EMAIL_KEY` rather than the `SECRET_KEY`
in an applications config. This is because
- this will be consistent with the other email tokens we generate
- although there is technically no harm in using `SECRET_KEY`,
the actual purpose of the `SECRET_KEY` is to be used for encrypting
Flask sessions. This would mean if we needed to rotate a key for
our Flask sessions it would also break out email tokens and vice
versa. By using different keys we should reduce the risk of a
token rotation having an impact on too large a part of the site

We maintain backwards compatability so tokens generated within the
past day using the old token will not be broken by this change.
The next step will be to start encrypting tokens with
`SHARED_EMAIL_KEY` for a full day at which point we can remove
the backwards compatability support for `SECRET_KEY`.