request id handling: back-populate the first value of DM_TRACE_ID_HEADERS back to the DM_REQUEST_ID_HEADER setting
because it turns out that some components (notably the apiclient) depend on that
setting - we can get rid of this once that/those are fixed. this also has
the nice property that DM_REQUEST_ID_HEADER should now always equal the
first entry of DM_TRACE_ID_HEADERS, even if it was DM_TRACE_ID_HEADERS itself
that was overridden.