Centralised app _status endpoint
Brings the core logic for serving an app's `_status` endpoint into utils
to share code and reduce replication. Each app should now need to only
call `get_app_status` and pass in the appropriate arguments. The default
status check will validate remaining disk space on the instance as well
as connectivity with the data/search API if applicable. If extra checks
need to be performed, these can be provided using `additional_checks`,
which is a sequence of callables that return dicts with the data that
should be included in the JSON response.

We also add a check for disk usage into the centralised status endpoint.
If disk space is below 5% (by default), we will return a 500 HTTP status
code causing Cloud Foundry to mark the instance as unhealthy and recycle
it.