Remove Python2 compatibility
Bump to 35.0.0 as we are dropping Python3 compatibility.

We've done a lot of work migrating our apps and libs to Python 3, so we
are no longer actively looking to support Python 2. The only known
things left running on Python 2 are a couple of scripts in dm-scripts,
but we have tickets to update these to Python 3 in the immediate future
and they are unlikely to be run before that happens.