Add framework datetime formatter
 ## Summary
We are going to store important framework lifecycle datetimes on the
API, so we need to be able to format these correctly for display on the
frontend. We currently use a different display format for these, so this
adds a formatter that will give the desired display e.g. '5pm BST,
Monday 1 July 2010'.

Also moves some formatting that takes place on datetime output to bring
us inline with the style guide to a more central method.

 ## Ticket
https://trello.com/c/BVPgZRZc/65-store-datetimes-relating-to-the-framework-lifecycle-on-the-api