Move api stubs here from dm-apiclient
 ## Summary
API stubs used to be defined in the apiclient, however they now require
some formatting that's held in this repository. Rather than introduce an
extra dependency to the apiclient, which would likely introduce
dependency conflicts when both libs are required by a frontend app,
we'll move the stubs to utils.

 ## Ticket
https://trello.com/c/BVPgZRZc/65-store-datetimes-relating-to-the-framework-lifecycle-on-the-api