Add monthyear format
 ## Summary
Add a date formatter that turns a datetime object/string into a month
and year for displaying to users when we don't care about precision - we
just want to convey roughly when an event occurred. We use this
currently to display when a supplier made a previous application to a
framework to remind them roughly when they gave answers/when the answers
were valid.