Log duration of calls to S3
 ## Summary
We think we have some performance issues with calls to S3 on certain
views. This replaces the existing logs for these calls and inserts the
duration of the call so that we can start to investigate the issue more
thoroughly.

It also adds logs/durations to any calls out to S3 that were missing
them, with a more restrictive condition that the call must be 'slow' or
sampled.

This PR also defines a 'slow' request to an external provider as one
that takes 250ms or longer.