Log duration of email calls
 ## Summary
We now log the duration of calls to AWS S3, but we should expand this to
our other third party providers. This PR introduces `logged_duration`
wrappers around the main calls to Mailchimp, Mandrill, and Notify. It
also introduces a generic partial that can be used to quickly wrap calls
to third-party providers and emit a consistent log entry, which involved
refactoring the current S3 logged_duration calls slightly.

 ## Ticket
https://trello.com/c/M0hAGLYT/303