Add a wtform error formatter
 ## Summary
We want to format errors with a standard interface so that we can pass
them into templates and always interact with them in the same way,
enhancing the capabilities of our templates to react to errors in a
consistent way. This adds a method `dmutils.forms.format_wtform_errors`
for parsing and formatting errors from Flask-WTForms into the same
format generated by content-loader form errors.

 ## Ticket
https://trello.com/c/RSYEM3FL/448