Do not configure urllib3 retry logger for search-api
The search-api uses the Elasticsearch client. The client uses urllib3 to
make requests, but has its own logic for managing retries. When the
client sends a request via urllib3 it always sets retries to `False`.
This causes urllib3.util.retry to emit an unwanted log message, twice,
for every request. It looks like this: `Converted retries value: False
-> Retry(total=False, connect=None, read=None, redirect=0, status=None)`
.

As retries are not configured via our apiclient, and there is no way to
set the retry config for urllib3 when used by Elasticsearch, we can
safely drop the logger for the search-api.