Refactor api stub tests
The api stubs have gotten stale, and part of the reason I think is
because the test suites makes updating them more effortful.

This commit refactors the tests to be more orthogonal, so that at only
one or at most two tests need to be updated when the api stubs get
changed.