boto: bump boto3, botocore & moto dependencies
this is really to allow us to have a newer moto (in this & other
repositories) which supports recent aws features. the botocore<1.11.0
stipulation is in place to ensure we end up using a version of
botocore in production that is compatible with the version we're
testing against, which is restricted by moto having an incompatibility
with botocore >= 1.11.0 (see https://github.com/spulec/moto/issues/1793).