Make the dm_mandrill client a bit more ergonomic
Breaking change.

Although Mandrill is deprecated and only used in one app these days, I
refactored it because it is still required in one app
(supplier-frontend) and was a pain to use.

This commit makes it more in line with the DMNotifyClient, by making the
client wrapper a class, and also it makes the order of arguments to
send_email a bit more intuitive (IMO).