Remove pinning from pytz package
The pytz package provides a database of timezones that is constantly
being updated, which can be used via a fairly simple and stable API.

I think that given this it doesn't make much sense to pin a version or
range of versions; given that ignoring updates can lead to timezone bugs
which are always tricky to debug, and there is little chance of breakage
due to an API change.

Of course, apps that consume the dmutils library can always choose their
own policy and pin a specific version of pytz.