Fix issue with pytest>=3.3.0
In pytest 3.3.0 default logging capture handlers are added to the
logging configuraton.

This interferes with our own logging capture fixtures. In the long term
we should look at changing our tests to use the pytest built in
capabilities, but for now we can disable the new features with this
patch.