Add reply_to email address parameter to send_email method
I've added a way for users of this API to send an email with a reply-to address. This is capability that has been added to the Notify library, which this wraps. To use the reply-to email, it has to first be set up with Notify. It's really only suitable for reusing a common reply-to email address, but luckily that's our only use case for the moment.