Reorder requirements
This commit exists to mitigate a problem with the libraries not pinning to a specific version. The version of attrs installed for hypothesis is high enough to clear the bar of pytests requirements but not actually high enough. Laurence's solution, to reorder the libraries to ensure pytests requirements are met first, works beautifully.