logged_duration default condition: add a moderately large SLOW_DEFAULT_CALL_THRESHOLD
the thinking here is that *any* routine we consider measurable spending
more than half a second doing what it's doing is probably something we
want to know about and the maximum rate it can hypothetically churn our
unwanted logs is 2 per second (per thread I guess). longer than
SLOW_EXTERNAL_CALL_THRESHOLD because we already try and limit the number
of external calls we do, whereas we may not have the same attitude to
whatever block of code this is wrapped around.

a logged_duration user wanting to use the default condition but a
shorter threshold can always add that to the default condition with
an "or" clause