Add params for GOV.UK Frontend error messages and summaries to WTForms error helper
We want to be able to use the GOV.UK Frontend error components in our
frontend apps.

This commit adds the expected parameters for the govuk-frontend error
summary and error message components to the output of
`get_errors_from_wtform`, so that you can create an error summary in
GOV.UK Design System styles from a form validated by WTForms.

This change should be backwards compatible.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>