Block ')' from ending URLs
Previously, the URL-detection heuristic would see "something (like http://example.com) this" and detect "http://example.com)" as the URL. Whilst URLs can validly include ')', it is very rare. Indeed, of the 1000+ closed briefs we have, 49 have URLs that end in ')' where they should not. No closed brief has a URL that intentionally ends in ')'.

You can see an example of this problem in https://www.digitalmarketplace.service.gov.uk/digital-outcomes-and-specialists/opportunities/13014, which was reported in x-gov Slack.

Thus, update our URL-detection heuristic to to forbid ')' as the last character of a URL. This will do the right thing for users in the vast majority of cases. I have not found any URLs that this change will break.

Add tests to confirm the new behaviour.